refactor(trashBin): group tasks by object type for better organization (#5660)
### 📣 Summary
Split trash bin tasks into separate files based on the object they
operate on.



### 📖 Description
Previously, all trash bin–related tasks were located in a single
tasks.py file, regardless of the object type (e.g., projects, accounts).
This refactor organizes tasks by grouping them into separate files
according to the object they manage.

This restructuring sets the stage for integrating attachment-related
tasks cleanly and consistently.
