+++ kobo/apps/trash_bin/tests/test_utils.py
@@ -1,5 +1,4 @@
 from datetime import timedelta
-from unittest.mock import patch
 
 from django.conf import settings
 from django.contrib.auth import get_user_model
@@ -56,9 +55,7 @@ class AccountTrashTestCase(TestCase):
             retain_placeholder=False,
         )
         account_trash = AccountTrash.objects.get(user=someuser)
-        with patch('kobo.apps.trash_bin.tasks.delete_kc_user') as mock_delete_kc_user:
-            mock_delete_kc_user.return_value = True
-            empty_account.apply([account_trash.pk])
+        empty_account.apply([account_trash.pk])
 
         assert not get_user_model().objects.filter(pk=someuser_id).exists()
         assert not Asset.objects.filter(owner_id=someuser_id).exists()
@@ -180,9 +177,7 @@ class AccountTrashTestCase(TestCase):
             retain_placeholder=True,
         )
         account_trash = AccountTrash.objects.get(user=someuser)
-        with patch('kobo.apps.trash_bin.tasks.delete_kc_user') as mock_delete_kc_user:
-            mock_delete_kc_user.return_value = True
-            empty_account.apply([account_trash.pk])
+        empty_account.apply([account_trash.pk])
         after = now() + timedelta(days=grace_period)
 
         someuser.refresh_from_db()
